requirements.txt
six==1.12.0
sympy==1.3
cryptography==3.3.2
urllib3==1.24.2
requests==2.20.0
gmpy==1.17
gmpy2==2.1.0b5
pycryptodome==3.9.7
tqdm
z3-solver
